Moved profile components into a Profile folder
